refactor(app): перенес компонент страницы продуктов в одноименный каталог
